+++ model_api/python/openvino/model_api/tilers/instance_segmentation.py
@@ -141,7 +141,7 @@ class InstanceSegmentationTiler(DetectionTiler):
         for i in range(detections_array.shape[0]):
             label = int(detections_array[i][0])
             score = float(detections_array[i][1])
-            bbox = list(detections_array[i][2:])
+            bbox = list(detections_array[i][2:].astype(np.int32))
             masks[i] = _segm_postprocess(np.array(bbox), masks[i], *shape[:-1])
             detected_objects.append(
                 SegmentedObject(*bbox, score, label, self.model.labels[label], masks[i])
